Fixes
fixede some broken functionality like words, vowels and consonants not getting the correct values to the right place. made textchanger more generic to allow the same function to work for more pages.

planning to clean up some code, get rid of duplicates and implimenting enums to make it easier to work with lists and be less dependend on I